// src/client/stores/ActionType.ts
export enum ActionType {
    UPDATE_USER = 'UPDATE_USER',
    UPDATE_PASSWORD = 'UPDATE_USER',

    LOGIN = 'LOGIN',

    SIGNOUT = 'SIGNOUT',

    SIGNUP_REQUEST = 'SIGNUP_REQUEST',
    SIGNUP_SUCCESS = 'SIGNUP_SUCCESS',
    SIGNUP_FAILURE = 'SIGNUP_FAILURE',

    LOGIN_FACEBOOK_REQUEST = 'LOGIN_FACEBOOK_REQUEST',
    LOGIN_FACEBOOK_SUCCESS = 'LOGIN_FACEBOOK_SUCCESS',
    LOGIN_FACEBOOK_FAILURE = 'LOGIN_FACEBOOK_FAILURE',

    LOGIN_REQUEST = 'LOGIN_REQUEST',
    LOGIN_SUCCESS = 'LOGIN_SUCCESS',
    LOGIN_FAILURE = 'LOGIN_FAILURE',

    GET_USER_REQUEST = 'GET_USER_REQUEST',
    GET_USER_SUCCESS = 'GET_USER_SUCCESS',
    GET_USER_FAILURE = 'GET_USER_FAILURE',

    UPDATE_USER_REQUEST = 'UPDATE_USER_REQUEST',
    UPDATE_USER_SUCCESS = 'UPDATE_USER_SUCCESS',
    UPDATE_USER_FAILURE = 'UPDATE_USER_FAILURE',

    UPDATE_PASSWORD_REQUEST = 'UPDATE_PASSWORD_REQUEST',
    UPDATE_PASSWORD_SUCCESS = 'UPDATE_PASSWORD_SUCCESS',
    UPDATE_PASSWORD_FAILURE = 'UPDATE_PASSWORD_FAILURE',

    GET_GAME_REQUEST = 'GET_GAME_REQUEST',
    GET_GAME_SUCCESS = 'GET_GAME_SUCCESS',
    GET_GAME_FAILURE = 'GET_GAME_FAILURE',

    CLEAR_ERRORS = 'CLEAR_ERRORS'
}